organization of the code in functions
